Moved the functions starting with check to a separate file
