+++ compiler/rustc_typeck/src/check/mod.rs
@@ -76,6 +76,7 @@ pub mod dropck;
 mod expectation;
 mod expr;
 mod fn_ctxt;
+mod fn_ctxt_checks;
 mod fn_ctxt_impl;
 mod fn_ctxt_suggestions;
 mod gather_locals;
@@ -99,6 +100,7 @@ pub use check::{check_item_type, check_wf_new};
 pub use diverges::Diverges;
 pub use expectation::Expectation;
 pub use fn_ctxt::FnCtxt;
+pub use fn_ctxt_checks::*;
 pub use fn_ctxt_impl::*;
 pub use fn_ctxt_suggestions::*;
 pub use inherited::{Inherited, InheritedBuilder};
